Change logic to be membership check
