+++ association-backend/src/webhooks/handlers/__init__.py
@@ -3,7 +3,7 @@ from typing import Any, Callable, Literal, Optional
 
 from src.webhooks.exceptions import WebhookError
 
-from .crons.membership_check_expiration import membership_check_expiration
+from .crons.membership_check_status import membership_check_status
 from .pretix.pretix_event_order_paid import pretix_event_order_paid
 from .stripe.handle_invoice_paid import handle_invoice_paid
 
@@ -15,7 +15,7 @@ HANDLERS = {
         "invoice.paid": handle_invoice_paid,
     },
     "pretix": {"pretix.event.order.paid": pretix_event_order_paid},
-    "crons": {"membership.check_expiration": membership_check_expiration},
+    "crons": {"membership.check_status": membership_check_status},
 }
