--- association-backend/src/webhooks/handlers/crons/membership_check_expiration.py
@@ -1,5 +1,0 @@
-from typing import Any
-
-
-async def membership_check_expiration(payload: Any):
-    pass
