+++ association-backend/src/webhooks/handlers/crons/membership_check_status.py
@@ -1,0 +1,6 @@
+from typing import Any
+
+
+async def membership_check_status(payload: Any):
+    print("abc")
+    pass
